+++ packages/synapse-interface/components/Maintenance/Events/example/EcotoneForkUpgrade.tsx
@@ -1,6 +1,6 @@
 import { useBridgeState } from '@/slices/bridge/hooks'
 import { useIntervalTimer } from '@/utils/hooks/useIntervalTimer'
-import { METIS } from '@/constants/chains/master'
+import { OPTIMISM, BASE } from '@/constants/chains/master'
 import {
   useEventCountdownProgressBar,
   getCountdownTimeStatus,
@@ -19,10 +19,12 @@ import { WarningMessage } from '../../../Warning'
  * End: 50 min after start of Ecotone Fork Upgrade Time
  */
 export const ECOTONE_FORK_BANNERS_START = new Date(
-  Date.UTC(2024, 2, 20, 16, 0, 0)
+  Date.UTC(2024, 2, 13, 23, 20, 0)
 )
-export const ECOTONE_FORK_START_DATE = new Date(Date.UTC(2024, 2, 20, 18, 0, 0))
-export const ECOTONE_FORK_END_DATE = new Date(Date.UTC(2024, 2, 20, 21, 0, 0))
+export const ECOTONE_FORK_START_DATE = new Date(
+  Date.UTC(2024, 2, 13, 23, 35, 0)
+)
+export const ECOTONE_FORK_END_DATE = new Date(Date.UTC(2024, 2, 14, 0, 25, 0))
 
 /** Previous implementation can be seen here: https://github.com/synapsecns/sanguine/pull/2294/files#diff-bbe6298d3dfbc80e46e2ff8b399a3e1822cede80f392b1af91875145ad4eeb19R19 */
 export const EcotoneForkUpgradeBanner = () => {
@@ -35,9 +37,13 @@ export const EcotoneForkUpgradeBanner = () => {
 
   return (
     <AnnouncementBanner
-      bannerId="03202024-metis-pause"
+      bannerId="03142024-ecotone-fork"
       bannerContents={
-        <>Metis Bridging is paused. Will be back online shortly.</>
+        <>
+          Optimism + Base Bridging will be paused 10 minutes ahead of Ecotone
+          (March 14, 00:00 UTC, 20:00 EST). Will be back online shortly
+          following the network upgrade.
+        </>
       }
       startDate={ECOTONE_FORK_BANNERS_START}
       endDate={ECOTONE_FORK_END_DATE}
@@ -54,14 +60,18 @@ export const EcotoneForkUpgradeBanner = () => {
 export const EcotoneForkWarningMessage = () => {
   const { fromChainId, toChainId } = useBridgeState()
 
-  const isToChainMetis = toChainId === METIS.id
+  const isChainOptimism = [fromChainId, toChainId].includes(OPTIMISM.id)
+  const isChainBase = [fromChainId, toChainId].includes(BASE.id)
 
-  if (isToChainMetis) {
+  if (isChainOptimism || isChainBase) {
     return (
       <WarningMessage
         message={
           <>
-            <p>Metis Chain bridging are paused.</p>
+            <p>
+              Optimism Chain and Base Chain bridging are paused until the
+              Ecotone Fork upgrade completes.
+            </p>
           </>
         }
       />
@@ -95,22 +105,23 @@ export const EcotoneForkWarningMessage = () => {
 export const useEcotoneForkCountdownProgress = () => {
   const { fromChainId, toChainId } = useBridgeState()
 
-  const isToChainMetis = toChainId === METIS.id
+  const isChainOptimism = [fromChainId, toChainId].includes(OPTIMISM.id)
+  const isChainBase = [fromChainId, toChainId].includes(BASE.id)
 
   const {
     isPending: isEcotoneForkUpgradePending,
     EventCountdownProgressBar: EcotoneForkCountdownProgressBar,
   } = useEventCountdownProgressBar(
-    'Metis Bridging is paused',
+    'Ecotone Fork upgrade in progress',
     ECOTONE_FORK_START_DATE, // Countdown Bar will automatically appear after start time
     ECOTONE_FORK_END_DATE // Countdown Bar will automatically disappear when end time is reached
   )
 
   return {
     isEcotoneForkUpgradePending,
-    isCurrentChainDisabled: isToChainMetis, // Used to pause Bridge
-    EcotoneForkCountdownProgressBar: isToChainMetis
-      ? EcotoneForkCountdownProgressBar
-      : null,
+    isCurrentChainDisabled:
+      (isChainOptimism || isChainBase) && isEcotoneForkUpgradePending, // Used to pause Bridge
+    EcotoneForkCountdownProgressBar:
+      isChainOptimism || isChainBase ? EcotoneForkCountdownProgressBar : null,
   }
 }